# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface FaqSummary {
  slug: string;
  question: string;
  shortAnswer: string;
  categorySlug: string;
  categoryTitle: string;
  topicSlugs: string[];
  tags: string[];
}

export interface FaqDetail {
  slug: string;
  question: string;
  shortAnswer: string;
  answer: string;
  answerHtml: string | null;
  metaDescription: string;
  categorySlug: string;
  categoryTitle: string;
  topicSlugs: string[];
  tags: string[];
  relatedServiceSlugs: string[];
  relatedFaqSlugs: string[];
  pillarBlogSlug: string | null;
  updatedDate: string;
}

export interface FaqListResponse {
  items: FaqSummary[];
  total: number;
}

export interface FaqDetailResponse {
  faq: FaqDetail;
  related: FaqSummary[];
}

export interface DuplicateMatch {
  slug: string;
  question: string;
  score: number;
}

export interface DuplicateCheckResponse {
  query: string;
  matches: DuplicateMatch[];
}

export type ListFaqsParams = {
category?: string;
topic?: string;
service?: string;
q?: string;
featured?: boolean;
};

export type CheckFaqDuplicateParams = {
q: string;
};