# lib/api-zod/src/generated/types/faqDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FaqDetail {
  slug: string;
  question: string;
  shortAnswer: string;
  answer: string;
  answerHtml: string | null;
  metaDescription: string;
  categorySlug: string;
  categoryTitle: string;
  topicSlugs: string[];
  tags: string[];
  relatedServiceSlugs: string[];
  relatedFaqSlugs: string[];
  pillarBlogSlug: string | null;
  updatedDate: string;
}