# lib/api-zod/src/generated/types/faqDetailResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FaqDetail } from './faqDetail';
import type { FaqSummary } from './faqSummary';

export interface FaqDetailResponse {
  faq: FaqDetail;
  related: FaqSummary[];
}