# lib/api-zod/src/generated/types/faqSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FaqSummary {
  slug: string;
  question: string;
  shortAnswer: string;
  categorySlug: string;
  categoryTitle: string;
  topicSlugs: string[];
  tags: string[];
}